backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.45
pydantic==2.9.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.12.1
redis==5.0.1
pymysql==1.1.0
requests==2.31.0
aiofiles==25.1.0
email-validator==2.1.0.post1